Trying to add ccminer bin mining service
