+++ app/src/main/AndroidManifest.xml
@@ -74,7 +74,7 @@
         </activity>
 
         <service
-            android:name="shmutalov.verusminer9000.miner.ScalaMiningService"
+            android:name="shmutalov.verusminer9000.miner.VerusBinMiningService"
             android:enabled="true"/>
 
         <receiver
